chore: move host kata dump into kata.go
